The most significant changes involve the addition of new files `hashTableLinear.cpp` and `hashTableLinear.h` which introduce a new class `HashTableLinear`. This class includes methods for resizing the hash table, inserting and removing elements, and searching for a key. It also includes a destructor and a constructor that initializes the hash table with a given capacity. Additionally, the `resize` function's parameter type in `hashTableChaining.cpp` and `hashTableChaining.h` was changed from `int` to `size_t` to handle larger sizes.
List of changes:

1. The `arrayMalloc.vcxproj` and `arrayMalloc.vcxproj.filters` files were updated to include the new `hashTableLinear.cpp` and `hashTableLinear.h` files. This ensures these new files will be compiled and included in the project.

2. The `resize` function's parameter type was changed from `int` to `size_t` in `hashTableChaining.cpp` and `hashTableChaining.h`. This change allows the function to handle larger sizes.

3. A new file `hashTableLinear.cpp` was added. This file includes the implementation of a new class `HashTableLinear` with methods for resizing the hash table, inserting and removing elements, and searching for a key. It also includes a destructor and a constructor.

4. A new file `hashTableLinear.h` was added. This file includes the declaration of the `HashTableLinear` class and its methods.

5. The `getchar();` line was removed from `main.cpp`. This line was likely used for debugging purposes, and its removal suggests that the debugging process is complete.

+++ hashTableChaining.cpp
@@ -1,6 +1,6 @@
 #include "hashTableChaining.h"
 
-void HashTableChaining::resize(int newCapacity) {
+void HashTableChaining::resize(size_t newCapacity) {
 	Node** newTable = new Node * [newCapacity];
 	for (size_t i = 0; i < newCapacity; i++) {
 		newTable[i] = NULL;
